Refactor: Update imports and add logout functionality in Header component
Update imports in the Header component to remove unnecessary type casting and improve code readability. Additionally, add logout functionality to allow users to log out of their accounts.

Update responsive styles for cards and proposal details

Improve the responsive styles for cards and proposal details to enhance the user experience on different devices.

Add useCheckDelegationChains hook

Create a new hook called useCheckDelegationChains to check for delegation chains in the application. This hook verifies the presence of a specific key in the localStorage and removes the 'user' key if the delegation chains are empty.

Update ProposalDetail component

Replace the base64 image source with the direct URL in the ProposalDetail component to improve performance and reduce unnecessary data processing.